requirements.txt
GDAL==3.4.1
geopy==2.4.0
numpy==1.23.3
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==10.1.0
pymongo==4.6.0
python-dotenv==1.0.0
streamlit==1.28.0
ultralytics==8.0.202
